+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### **Added**
 
+- added VPC/subnets/sg config for multi-account project template to `sagemaker-templates-service-catalog` module
 - added `sagemaker-custom-kernel` module
 - added batch inference project template to `sagemaker-templates-service-catalog` module
 - added EFS removal policy to `mlflow-fargate` module